Eval

Documentation

applyOp :: Op -> Exp -> Exp -> Exp

Apply a binary operation after evaluating two expressions

eval :: Exp -> Exp

Evaluate an expression via substitution

substitute :: Exp -> String -> Exp -> Exp

Substitute an expression for a variable name in an expression

getIds :: Exp -> [String]

Get a list of variable names in an expression. I should get only free variables.

evalMain :: [Assign] -> Exp

Given a set of assignments, evaluate the expression named main